+++ EstimoteSDK/EstimoteSDK.framework/Versions/A/Headers/EstimoteSDK.h
@@ -7,7 +7,7 @@
 //  |______|___/\__|_|_| |_| |_|\___/ \__\___| |_____/|_____/|_|\_\
 //
 //
-//  Version: 4.26.1
+//  Version: 4.26.2
 //  Copyright (c) 2016 Estimote. All rights reserved.
 
 /**